test: cover FP32-to-INT8 and INT32-to-FP32
