+++ examples/balls/fp2int/verify/src/agents/mem/fp2int_mem_model.svh
@@ -4,7 +4,6 @@ class fp2int_mem_model extends uvm_component;
   virtual fp2int_if vif;
   uvm_analysis_imp_stim #(fp2int_cmd_item, fp2int_mem_model) stim_imp;
   bit [127:0] input_words[FP2INT_MAX_WORDS];
-  bit is_i8;
   bit have_stim;
 
   function new(string name, uvm_component parent);
@@ -23,7 +22,6 @@ class fp2int_mem_model extends uvm_component;
     for (int i = 0; i < FP2INT_MAX_WORDS; i++) begin
       input_words[i] = item.input_words[i];
     end
-    is_i8 = item.is_i8();
     have_stim = 1'b1;
   endfunction
 
@@ -70,15 +68,7 @@ class fp2int_mem_model extends uvm_component;
         `uvm_fatal("MEM", $sformatf("read addr out of range: %0d", vif.bank_read_0_req_bits_addr))
       end
 
-      if (is_i8) begin
-        word_index = vif.bank_read_0_req_bits_addr * FP2INT_NUM_GROUPS + vif.bank_read_0_group_id;
-      end else begin
-        if (vif.bank_read_0_group_id !== 5'd0) begin
-          `uvm_fatal("MEM", $sformatf("unexpected group_id for INT32 mode: %0d",
-                                      vif.bank_read_0_group_id))
-        end
-        word_index = vif.bank_read_0_req_bits_addr;
-      end
+      word_index = vif.bank_read_0_req_bits_addr * FP2INT_NUM_GROUPS + vif.bank_read_0_group_id;
 
       vif.bank_read_0_resp_bits_data <= input_words[word_index];
       vif.bank_read_0_resp_valid <= 1'b1;
