+++ examples/balls/fp2int/workloads/ctests/CMakeLists.txt
@@ -1,6 +1,3 @@
 add_buckyball_ctests(
-  fp2int_fp32_to_int32_test.c
   fp2int_fp32_to_int8_test.c
-  fp2int_bank_fp32_to_int32_test.c
-  fp2int_bank_fp32_to_int8_test.c
 )
