+++ examples/balls/fp2int/workloads/mlir_tests/ball/fp2int_2x16_contract.mlir
@@ -4,9 +4,9 @@
 
 func.func @main() -> i8 {
   %z = arith.constant 0 : i8
-  %iter = arith.constant 4 : i64
-  %scale = arith.constant 1065353216 : i64
-  %in = buckyball.bank_alloc
+  %iter = arith.constant 2 : i64
+  %scale = arith.constant 1073741824 : i64
+  %in = buckyball.bank_alloc {col = 4 : i64}
   %out = buckyball.bank_alloc
   buckyball.fp2int %in, %out, %iter, %scale : i64
   buckyball.bank_release %in : i64
