+++ examples/balls/fp2int/workloads/mlir_tests/bank/CMakeLists.txt
@@ -3,7 +3,7 @@ include(${OPTEST_WORKLOAD_DIR}/cmake/BuckyballMlirTests.cmake)
 set(BUCKYBALL_MLIR_GROUP_TARGET balls-mlir-tests-build)
 set(BUCKYBALL_MLIR_TEST_PREFIX bank)
 
-add_buckyball_mlir_contract(fp2int_4x4_contract
+add_buckyball_mlir_contract(fp2int_2x16_contract
   PASSES "--assign-physical-banks=bank_num=${BUCKYBALL_MLIR_BANK_NUM}")
 
-add_buckyball_mlir_optest(fp2int_4x4)
+add_buckyball_mlir_optest(fp2int_2x16)
