+++ examples/balls/fp2int/workloads/mlir_tests/bank/fp2int_2x16_main.cpp
@@ -2,8 +2,9 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-static const int32_t expected[16] = {
-    1, 2, 3, -1, -2, 0, 4, 5, 10, -10, 0, 100, -100, 7, 8, -8,
+static const int8_t expected[32] = {
+    0, 0,  0, 0,  2, -2, 2, -2, 4,  -4,  126, 127, -128, -128, 0,   0,
+    4, -4, 6, -6, 6, -6, 8, -8, 20, -20, 40,  -40, 1,    -1,   127, -128,
 };
 
 static void fail(void) {
@@ -20,21 +21,21 @@ static void fail(void) {
 #ifdef __cplusplus
 extern "C"
 #endif
-    void check_result(int32_t *allocated, int32_t *aligned, int64_t offset,
+    void check_result(int8_t *allocated, int8_t *aligned, int64_t offset,
                       int64_t size0, int64_t size1, int64_t stride0,
                       int64_t stride1) {
   (void)allocated;
-  if (size0 != 4 || size1 != 4 || stride0 != 4 || stride1 != 1) {
+  if (size0 != 2 || size1 != 16 || stride0 != 16 || stride1 != 1) {
     printf("FAILED: bank_fp2int shape %dx%d stride %dx%d\n", (int)size0,
            (int)size1, (int)stride0, (int)stride1);
     fail();
   }
 
-  int32_t *out = aligned + offset;
-  for (int i = 0; i < 4; ++i) {
-    for (int j = 0; j < 4; ++j) {
-      int32_t exp = expected[i * 4 + j];
-      int32_t got = out[i * stride0 + j * stride1];
+  int8_t *out = aligned + offset;
+  for (int i = 0; i < 2; ++i) {
+    for (int j = 0; j < 16; ++j) {
+      int8_t exp = expected[i * 16 + j];
+      int8_t got = out[i * stride0 + j * stride1];
       if (got != exp) {
         printf("FAILED: bank_fp2int out[%d][%d] exp=%d got=%d\n", i, j,
                (int)exp, (int)got);
@@ -42,5 +43,5 @@ extern "C"
       }
     }
   }
-  printf("PASSED: fp2int bank_fp2int 4x4 fp32->int32\n");
+  printf("PASSED: fp2int bank_fp2int 2x16 fp32->int8\n");
 }
