+++ examples/balls/int2fp/workloads/ctests/CMakeLists.txt
@@ -1,12 +1,3 @@
 add_buckyball_ctests(
   int2fp_int32_to_fp32_test.c
-  int2fp_int8_to_fp32_test.c
-  int2fp_int32_to_int8_test.c
-  int2fp_int32_to_fp32_col2_test.c
-  int2fp_int32_to_fp32_col4_test.c
-  int2fp_bank_int32_to_fp32_test.c
-  int2fp_bank_int8_to_fp32_test.c
-  int2fp_bank_int32_to_int8_test.c
-  int2fp_bank_int32_to_fp32_col2_test.c
-  int2fp_bank_int32_to_fp32_col4_test.c
 )
