+++ examples/balls/int2fp/workloads/ctests/int2fp_int32_to_fp32_test.c
@@ -3,70 +3,38 @@
 #include <bbhw/mem/mem.h>
 #include <stdint.h>
 #include <stdio.h>
-#include <stdlib.h>
+#define ROWS 2
+#define LANES 4
+#define SCALE_0_25 0x3E800000U
 
-#define DIM 16
-
-// INT32 input data (4 elements per SRAM word, 16 words)
-static int32_t int32_input[DIM * 4] __attribute__((aligned(64))) = {
-    1, 2, 3, -1, -2, 0, 4, 5, 10, -10, 7, 100, -100, 8, 16, -8,
-    1, 2, 3, -1, -2, 0, 4, 5, 10, -10, 7, 100, -100, 8, 16, -8,
-    1, 2, 3, -1, -2, 0, 4, 5, 10, -10, 7, 100, -100, 8, 16, -8,
-    1, 2, 3, -1, -2, 0, 4, 5, 10, -10, 7, 100, -100, 8, 16, -8,
+static int32_t int32_input[ROWS * LANES] __attribute__((aligned(64))) = {
+    1, -1, 2, -2, 3, -3, 5, -5,
 };
 
-// Expected FP32 output as bit patterns: int32_val * 1.0 = float(int32_val)
-// 1.0=0x3F800000, 2.0=0x40000000, 3.0=0x40400000, -1.0=0xBF800000
-// -2.0=0xC0000000, 0.0=0x00000000, 4.0=0x40800000, 5.0=0x40A00000
-// 10.0=0x41200000, -10.0=0xC1200000, 7.0=0x40E00000, 100.0=0x42C80000
-// -100.0=0xC2C80000, 8.0=0x41000000, 16.0=0x41800000, -8.0=0xC1000000
-static uint32_t expected_fp32[DIM * 4] __attribute__((aligned(64))) = {
-    0x3F800000, 0x40000000, 0x40400000, 0xBF800000, 0xC0000000, 0x00000000,
-    0x40800000, 0x40A00000, 0x41200000, 0xC1200000, 0x40E00000, 0x42C80000,
-    0xC2C80000, 0x41000000, 0x41800000, 0xC1000000, 0x3F800000, 0x40000000,
-    0x40400000, 0xBF800000, 0xC0000000, 0x00000000, 0x40800000, 0x40A00000,
-    0x41200000, 0xC1200000, 0x40E00000, 0x42C80000, 0xC2C80000, 0x41000000,
-    0x41800000, 0xC1000000, 0x3F800000, 0x40000000, 0x40400000, 0xBF800000,
-    0xC0000000, 0x00000000, 0x40800000, 0x40A00000, 0x41200000, 0xC1200000,
-    0x40E00000, 0x42C80000, 0xC2C80000, 0x41000000, 0x41800000, 0xC1000000,
-    0x3F800000, 0x40000000, 0x40400000, 0xBF800000, 0xC0000000, 0x00000000,
-    0x40800000, 0x40A00000, 0x41200000, 0xC1200000, 0x40E00000, 0x42C80000,
-    0xC2C80000, 0x41000000, 0x41800000, 0xC1000000,
+static uint32_t expected_fp32[ROWS * LANES] __attribute__((aligned(64))) = {
+    0x3E800000, 0xBE800000, 0x3F000000, 0xBF000000,
+    0x3F400000, 0xBF400000, 0x3FA00000, 0xBFA00000,
 };
 
-static uint32_t output_fp32[DIM * 4] __attribute__((aligned(64)));
+static uint32_t output_fp32[ROWS * LANES] __attribute__((aligned(64)));
 
-// FP32 bit pattern for scale = 1.0
-#define SCALE_1_0 0x3F800000U
-
-void hw_int2fp(int32_t *input, uint32_t *output, int num_words) {
-  uint32_t op1_bank_id = 0;
-  uint32_t wr_bank_id = 1;
-
-  bb_mem_alloc(op1_bank_id, 1, 1);
-  bb_mem_alloc(wr_bank_id, 1, 1);
-
-  bb_mvin((uintptr_t)input, op1_bank_id, num_words, 1);
-
-  bb_int2fp(op1_bank_id, wr_bank_id, num_words, SCALE_1_0);
-
-  bb_mvout((uintptr_t)output, wr_bank_id, num_words, 1);
-  bb_fence();
-}
-
-int main() {
+int main(void) {
 #ifdef MULTICORE
   multicore(MULTICORE);
 #endif
 
-  for (int i = 0; i < DIM * 4; i++) {
-    output_fp32[i] = 0;
-  }
+  const uint32_t src_bank = 0;
+  const uint32_t dst_bank = 1;
 
-  hw_int2fp(int32_input, output_fp32, DIM);
+  bb_mem_alloc(src_bank, ROWS, 1);
+  bb_mem_alloc(dst_bank, ROWS, 1);
+  bb_mvin((uintptr_t)int32_input, src_bank, ROWS, 1);
+  bb_int2fp(src_bank, dst_bank, ROWS, SCALE_0_25);
+  bb_mvout((uintptr_t)output_fp32, dst_bank, ROWS, 1);
+  bb_fence();
 
   int passed = 1;
-  for (int i = 0; i < DIM * 4; i++) {
+  for (int i = 0; i < ROWS * LANES; ++i) {
     if (output_fp32[i] != expected_fp32[i]) {
       printf("MISMATCH at [%d]: got 0x%08X, expected 0x%08X\n", i,
              output_fp32[i], expected_fp32[i]);
@@ -74,14 +42,6 @@ int main() {
     }
   }
 
-  if (passed) {
-    printf("Int2Fp INT32-to-FP32 test PASSED\n");
-  } else {
-    printf("Int2Fp INT32-to-FP32 test FAILED\n");
-  }
-  return (!passed);
-
-#ifdef MULTICORE
-  exit(0);
-#endif
+  printf("Int2Fp INT32-to-FP32 test %s\n", passed ? "PASSED" : "FAILED");
+  return passed ? 0 : 1;
 }
