+++ examples/balls/int2fp/workloads/mlir_tests/ball/CMakeLists.txt
@@ -10,4 +10,3 @@ add_buckyball_mlir_contract(int2fp_4x4_contract
     ${BUCKYBALL_LOWER_BUCKYBALL})
 
 add_buckyball_mlir_optest(int2fp_4x4)
-add_buckyball_mlir_optest(int32_to_int8_2x16)
