+++ examples/balls/int2fp/workloads/mlir_tests/bank/CMakeLists.txt
@@ -7,4 +7,3 @@ add_buckyball_mlir_contract(int2fp_4x4_contract
   PASSES "--assign-physical-banks=bank_num=${BUCKYBALL_MLIR_BANK_NUM}")
 
 add_buckyball_mlir_optest(int2fp_4x4)
-add_buckyball_mlir_optest(int32_to_int8_2x16)
