+++ examples/chips/pebble/regression/batch/bemu/workloads-pk.toml
@@ -36,12 +36,10 @@ tests = [
   "pebble_optest_tile_conv2d_6x6_k3-linux",
   "pebble_optest_bank_im2col_6x6_k3-linux",
   "pebble_optest_ball_im2col_6x6_k3-linux",
-  "pebble_optest_bank_fp2int_4x4-linux",
-  "pebble_optest_ball_fp2int_4x4-linux",
+  "pebble_optest_bank_fp2int_2x16-linux",
+  "pebble_optest_ball_fp2int_2x16-linux",
   "pebble_optest_bank_int2fp_4x4-linux",
   "pebble_optest_ball_int2fp_4x4-linux",
-  "pebble_optest_bank_int32_to_int8_2x16-linux",
-  "pebble_optest_ball_int32_to_int8_2x16-linux",
   "pebble_im2col_k1_test-linux",
   "pebble_im2col_k3_test-linux",
   "pebble_im2col_k3_partial_test-linux",
@@ -55,31 +53,15 @@ tests = [
   "pebble_conv_im2col_test-linux",
   "pebble_conv_transpose_test-linux",
   "pebble_conv_matmul_test-linux",
-  "pebble_quant_fp32_to_int8_test-linux",
   "pebble_quant_im2col_test-linux",
   "pebble_quant_transpose_test-linux",
   "pebble_quant_gemm_test-linux",
-  "pebble_quant_int8_to_fp32_test-linux",
   "pebble_conv_im2col_bank_test-linux",
   "pebble_conv_transpose_bank_test-linux",
   "pebble_conv_matmul_bank_test-linux",
-  "pebble_quant_fp32_to_int8_bank_test-linux",
   "pebble_quant_im2col_bank_test-linux",
   "pebble_quant_transpose_bank_test-linux",
   "pebble_quant_gemm_bank_test-linux",
-  "pebble_quant_int8_to_fp32_bank_test-linux",
-  "pebble_fp2int_fp32_to_int32_test-linux",
   "pebble_fp2int_fp32_to_int8_test-linux",
   "pebble_int2fp_int32_to_fp32_test-linux",
-  "pebble_int2fp_int8_to_fp32_test-linux",
-  "pebble_int2fp_int32_to_int8_test-linux",
-  "pebble_int2fp_int32_to_fp32_col2_test-linux",
-  "pebble_int2fp_int32_to_fp32_col4_test-linux",
-  "pebble_fp2int_bank_fp32_to_int32_test-linux",
-  "pebble_fp2int_bank_fp32_to_int8_test-linux",
-  "pebble_int2fp_bank_int32_to_fp32_test-linux",
-  "pebble_int2fp_bank_int8_to_fp32_test-linux",
-  "pebble_int2fp_bank_int32_to_int8_test-linux",
-  "pebble_int2fp_bank_int32_to_fp32_col2_test-linux",
-  "pebble_int2fp_bank_int32_to_fp32_col4_test-linux",
 ]
