+++ examples/chips/toy/regression/batch/bemu/workloads-elf.toml
@@ -56,20 +56,8 @@ tests = [
   "toy_gemmini_ws_risc_basic_test-singlecore-baremetal",
   "toy_mmio_minimal_test-singlecore-baremetal",
   "toy_mxfp2int_test-singlecore-baremetal",
-  "toy_fp2int_fp32_to_int32_test-singlecore-baremetal",
   "toy_fp2int_fp32_to_int8_test-singlecore-baremetal",
   "toy_int2fp_int32_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_int8_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_int8_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_fp32_col2_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_fp32_col4_test-singlecore-baremetal",
-  "toy_fp2int_bank_fp32_to_int32_test-singlecore-baremetal",
-  "toy_fp2int_bank_fp32_to_int8_test-singlecore-baremetal",
-  "toy_int2fp_bank_int32_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_bank_int8_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_bank_int32_to_int8_test-singlecore-baremetal",
-  "toy_int2fp_bank_int32_to_fp32_col2_test-singlecore-baremetal",
-  "toy_int2fp_bank_int32_to_fp32_col4_test-singlecore-baremetal",
   "toy_optest_buckyball_matmul_16x64_64x16_singlecore-baremetal",
   "toy_optest_buckyball_matmul_64x16_16x64_singlecore-baremetal",
   "toy_optest_buckyball_matmul_1024x16_16x1024_singlecore-baremetal",
