+++ examples/chips/toy/regression/batch/bemu/workloads-pk.toml
@@ -56,20 +56,8 @@ tests = [
   "toy_gemmini_ws_risc_basic_test-linux",
   "toy_mmio_minimal_test-linux",
   "toy_mxfp2int_test-linux",
-  "toy_fp2int_fp32_to_int32_test-linux",
   "toy_fp2int_fp32_to_int8_test-linux",
   "toy_int2fp_int32_to_fp32_test-linux",
-  "toy_int2fp_int8_to_fp32_test-linux",
-  "toy_int2fp_int32_to_int8_test-linux",
-  "toy_int2fp_int32_to_fp32_col2_test-linux",
-  "toy_int2fp_int32_to_fp32_col4_test-linux",
-  "toy_fp2int_bank_fp32_to_int32_test-linux",
-  "toy_fp2int_bank_fp32_to_int8_test-linux",
-  "toy_int2fp_bank_int32_to_fp32_test-linux",
-  "toy_int2fp_bank_int8_to_fp32_test-linux",
-  "toy_int2fp_bank_int32_to_int8_test-linux",
-  "toy_int2fp_bank_int32_to_fp32_col2_test-linux",
-  "toy_int2fp_bank_int32_to_fp32_col4_test-linux",
   "toy_optest_buckyball_matmul_16x64_64x16-linux",
   "toy_optest_buckyball_matmul_64x16_16x64-linux",
   "toy_optest_buckyball_matmul_1024x16_16x1024-linux",
