+++ examples/chips/toy/regression/batch/verilator/workloads-elf.toml
@@ -45,13 +45,8 @@ tests = [
   "toy_gemmini_ws_risc_basic_test-singlecore-baremetal",
   "toy_mmio_minimal_test-singlecore-baremetal",
   "toy_mxfp2int_test-singlecore-baremetal",
-  "toy_fp2int_fp32_to_int32_test-singlecore-baremetal",
   "toy_fp2int_fp32_to_int8_test-singlecore-baremetal",
   "toy_int2fp_int32_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_int8_to_fp32_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_int8_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_fp32_col2_test-singlecore-baremetal",
-  "toy_int2fp_int32_to_fp32_col4_test-singlecore-baremetal",
   "toy_tlb_bare_mode_test-singlecore-baremetal",
   "toy_tlb_sv39_identity_test-singlecore-baremetal",
   "toy_tlb_sfence_flush_test-singlecore-baremetal",
